requirements.txt
streamlit
pandas
openai-whisper
torch
pronouncing
ffmpeg-python